perf(grib): halve decoded grid buffers while preserving query precision (#751)
GRIB widened every decoded `f32` value to `f64` before caching it,
doubling the value buffers and evicting forecast fields sooner. Retain
the decoder's `f32` values and widen only sampled values before
interpolation. Coordinates, interpolation, display conversion and
EDR/Maps outputs retain their existing `f64` precision.

Cache weights now count allocated capacity plus key/structure overhead,
and the entry-size estimate follows the smaller grids. Adds a
reproducible release benchmark and regression coverage for cache
retention, allocation slack, interpolation precision, NaNs and signed
zero.

Measured with the committed 1440×721 ECMWF fixture, repeating the
comparison three times:

| Measurement | Before | After |
|---|---:|---:|
| Value-buffer bytes per field | 8,305,920 | 4,152,960 |
| Decodes for 36 requests over 12 message keys, 64 MiB cache | 21 | 12 |
| Resident grids after that replay | 8 | 12 |
| Median decode time | ~15.7 ms | ~15.5 ms |
| Warm 512×256 resampling | ~1.61 ms | ~1.66 ms |

The query-output fingerprint is identical before/after for point
sampling, area axes/values, and WGS84/Web Mercator/EPSG:3035 rendering.
The cache replay uses distinct keys with the same real fixture bytes; it
measures retention and decode counts, not live network latency. The
benefit is half-sized buffers and fewer repeated decodes; warm
resampling has a small conversion cost (~0.05 ms here).

Reproduce with `cargo run --release -p engine-grib --example
bench_grid_cache`.

Validation:

- `cargo fmt --all --check`
- `cargo test -p engine-grib -p api-edr --locked --offline`
- `cargo clippy --all-targets --locked --offline -- -D warnings`
- The small-cache regression fails on the previous implementation (three
decodes instead of one).

Related to #475; typed render-buffer variants and other engines remain
outside this change.

+++ crates/engine-grib/src/lib.rs
@@ -2022,7 +2022,7 @@ mod tests {
                     lat_first: 1.0,
                     lon_inc: 1.0,
                     lat_inc: -1.0,
-                    values: Arc::new(vec![(i + 1) as f64 * 10.0; 4]),
+                    values: Arc::new(vec![(i + 1) as f32 * 10.0; 4]),
                     triple: if i == 3 { (0, 4, 7) } else { (0, 1, 8) },
                     centre: 7,
                     first_surface_type: 1,
